requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain>=1.0.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Frontend
chainlit>=2.0.0

# LLM providers
google-genai>=1.0.0
mistralai>=1.0.0

# RAG
ragie>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
